Generic addon-a11y decorator
